% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.building.R
\name{lr.surv.pval}
\alias{lr.surv.pval}
\title{P values calculation in The multivariable Cox model Using LRT}
\usage{
lr.surv.pval(time, event, covariate, data)
}
\arguments{
\item{time}{A variable for time to event}

\item{event}{A variable for event indicator (1 indicates event, 0 is censored)}

\item{covariate}{A vector of variables included in the multivariable analysis}

\item{data}{A data set}
}
\value{
A matrix with variables and p values in the multivariable Cox model
}
\description{
The function creates a p-value table for multivariable Cox Proportional Hazard regression model.
}
\examples{
\dontrun{
library(survival)
df <- veteran
res <- lr.surv.pval(time="time", event="status",
            covariate=c("age","prior","trt","celltype","karno","diagtime"),
            data=df)
}
}
