% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.building.R
\name{surv.pval}
\alias{surv.pval}
\title{A backward elimination model building in Cox regression}
\usage{
surv.pval(time, event, covariate, sig = 0.05, data)
}
\arguments{
\item{time}{A variable for time to event}

\item{event}{A variable for event indicator (1 indicates event, 0 is censored)}

\item{covariate}{A vector of variables included in the multivariable analysis}

\item{sig}{A significance level that allows the variable in the finalized model}

\item{data}{A data set}
}
\value{
A matrix with variables and p values in the multivariable Cox model
}
\description{
The function builds or finalizes a multivariable Cox regression model based on a LRT p-value.
}
\examples{
\dontrun{
library(survival)
df <- veteran
res <- surv.pval(time="time", event="status",
            covariate=c("age","prior","trt","celltype","karno","diagtime"),
            sig=0.05, data=df)
}
}
